pyproject.toml
[project]
name = "chatbot-jne-simplificado"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "httpx>=0.28.1",
    "openai>=1.99.9",
    "oracledb>=3.3.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
